# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ZGRZYT API",
        "description": "Dokumentacja API dla systemu Zgłoszeń i Rejestru Zdarzeń Technicznych (ZGRZYT). Jest to webowa aplikacja do zgłaszania, obsługi i monitorowania błędów w systemach informatycznych.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/user": {
            "get": {
                "tags": [
                    "Użytkownicy"
                ],
                "summary": "Pobierz dane zalogowanego użytkownika",
                "description": "Zwraca dane użytkownika powiązanego z użytym tokenem API. Służy do weryfikacji tokena i pobrania podstawowych informacji (np. rola, nazwa). Jeśli token jest nieprawidłowy lub wygasł, serwer zwróci błąd 401 Unauthenticated.",
                "operationId": "getAuthenticatedUser",
                "responses": {
                    "200": {
                        "description": "Dane zalogowanego użytkownika.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Błąd autoryzacji (brak lub nieprawidłowy token)."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/users/{user}/activate": {
            "post": {
                "tags": [
                    "Użytkownicy"
                ],
                "summary": "Aktywacja konta użytkownika",
                "description": "Aktywuje nieaktywne konto użytkownika. Dostępne tylko dla ról 'admin' lub 'it'.",
                "operationId": "activateUser",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "ID użytkownika do aktywacji",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Konto pomyślnie aktywowane.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Konto użytkownika zostało pomyślnie aktywowane."
                                        },
                                        "user": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Brak uprawnień."
                    },
                    "404": {
                        "description": "Nie znaleziono użytkownika."
                    },
                    "409": {
                        "description": "Użytkownik jest już aktywny (Conflict)."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/users/{user}/ban": {
            "post": {
                "tags": [
                    "Użytkownicy"
                ],
                "summary": "Banowanie konta użytkownika",
                "description": "Banuje konto użytkownika. Dostępne tylko dla ról 'admin' lub 'it'.",
                "operationId": "banUser",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "ID użytkownika do zbanowania",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Konto pomyślnie zbanowane.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Konto użytkownika zostało pomyślnie zbanowane."
                                        },
                                        "user": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Brak uprawnień."
                    },
                    "404": {
                        "description": "Nie znaleziono użytkownika."
                    },
                    "409": {
                        "description": "Użytkownik jest już zbanowany (Conflict)."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/users/{user}/unban": {
            "post": {
                "tags": [
                    "Użytkownicy"
                ],
                "summary": "Odblokowanie (unban) konta użytkownika",
                "description": "Odblokowuje zbanowane konto użytkownika. Dostępne tylko dla ról 'admin' lub 'it'. Wymaga potwierdzenia hasłem osoby wykonującej akcję.",
                "operationId": "unbanUser",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "ID użytkownika do odbanowania",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Hasło użytkownika wykonującego operację w celu potwierdzenia.",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "password"
                                ],
                                "properties": {
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Konto pomyślnie odbanowane.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Konto użytkownika zostało pomyślnie odbanowane."
                                        },
                                        "user": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Brak uprawnień lub nieprawidłowe hasło."
                    },
                    "404": {
                        "description": "Nie znaleziono użytkownika."
                    },
                    "409": {
                        "description": "Użytkownik nie jest zbanowany (Conflict)."
                    },
                    "422": {
                        "description": "Błąd walidacji (np. brak hasła)."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "User": {
                "title": "User",
                "description": "Model użytkownika",
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": true,
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Jan Kowalski"
                    },
                    "login": {
                        "type": "string",
                        "example": "jkowalski"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "j.kowalski@example.com"
                    },
                    "role": {
                        "type": "string",
                        "example": "user",
                        "enum": [
                            "user",
                            "it",
                            "admin"
                        ]
                    },
                    "active": {
                        "description": "Status aktywności konta.",
                        "type": "boolean",
                        "example": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}